#331: Moved LB-related passive Parsers into the model folder
 Created a new "limitburt" folder to gather them.
