+++ src/app/ffbe/mappers/effects/passive-effect-parser.factory.ts
@@ -6,18 +6,18 @@ import {PassiveStatsIncreaseHpThresholdParser} from './passives/passive-stats-in
 import {PassiveEquipmentCategoryStatsIncreaseParser} from './passives/passive-equipment-category-stats-increase.parser';
 import {PassiveCoverParser} from './passives/passive-cover.parser';
 import {PassiveMpRecoveryParser} from './passives/passive-mp-recovery.parser';
-import {PassiveLbPerTurnParser} from './passives/passive-lb-per-turn.parser';
+import {PassiveLbPerTurnParser} from '../../model/effects/passives/limitburst/passive-lb-per-turn.parser';
 import {PassiveEvasionPhysicalParser} from './passives/passive-evasion-physical.parser';
-import {PassiveLbDamageIncreaseParser} from './passives/passive-lb-damage-increase.parser';
+import {PassiveLbDamageIncreaseParser} from '../../model/effects/passives/limitburst/passive-lb-damage-increase.parser';
 import {PassiveTargetChanceChangesParser} from './passives/passive-target-chance-changes.parser';
-import {PassiveLbSpeedIncreaseParser} from './passives/passive-lb-speed-increase.parser';
+import {PassiveLbSpeedIncreaseParser} from '../../model/effects/passives/limitburst/passive-lb-speed-increase.parser';
 import {PassiveDualWieldWeaponCategoryUnlockParser} from './passives/passive-dual-wield-weapon-category-unlock.parser';
 import {PassiveDebuffsResistanceParser} from './passives/passive-debuffs-resistance.parser';
 import {PassiveEsperGroupSummonParser} from './passives/passive-esper-group-summon.parser';
 import {PassiveJumpDamageIncreaseParser} from './passives/passive-jump-damage-increase.parser';
 import {PassiveSkillModifierIncreaseParser} from './passives/passive-skill-modifier-increase.parser';
-import {PassiveLbUpgradeParser} from './passives/passive-lb-upgrade.parser';
-import {PassiveLbUpgradeHpThresholdParser} from './passives/passive-lb-upgrade-hp-threshold.parser';
+import {PassiveLbUpgradeParser} from '../../model/effects/passives/limitburst/passive-lb-upgrade.parser';
+import {PassiveLbUpgradeHpThresholdParser} from '../../model/effects/passives/limitburst/passive-lb-upgrade-hp-threshold.parser';
 import {PassiveEsperStatsIncreaseParser} from './passives/passive-esper-stats-increase.parser';
 import {PassiveCounterAttackChanceParser} from './passives/passive-counter-attack-chance.parser';
 import {PassiveMpDecreaseForSongsParser} from './passives/passive-mp-decrease-for-songs.parser';
